test: Add a createStorage function to gmp/testing
Avoid code duplication for creating a storage object for testing
purposes.
