+++ src/gmp/__tests__/gmp.test.ts
@@ -18,16 +18,7 @@ import {
 } from 'gmp/commands/testing';
 import Gmp from 'gmp/gmp';
 import Settings from 'gmp/settings';
-
-const createStorage = (state?: Record<string, string>) => {
-  const store = {
-    state: {...state},
-    getItem: testing.fn(name => store.state[name] || null),
-    setItem: testing.fn((name, value) => (store.state[name] = String(value))),
-    removeItem: testing.fn(name => delete store.state[name]),
-  };
-  return store;
-};
+import {createStorage} from 'gmp/testing';
 
 let origLocation: Location;
