+++ src/gmp/session/__tests__/default-session.test.ts
@@ -6,16 +6,7 @@
 import {describe, test, expect, testing, onTestFinished} from '@gsa/testing';
 import date from 'gmp/models/date';
 import DefaultSession from 'gmp/session/default-session';
-
-const createStorage = (state?: Record<string, string>) => {
-  const store = {
-    state: {...state},
-    getItem: testing.fn(name => store.state[name] ?? null),
-    setItem: testing.fn((name, value) => (store.state[name] = String(value))),
-    removeItem: testing.fn(name => delete store.state[name]),
-  };
-  return store;
-};
+import {createStorage} from 'gmp/testing';
 
 describe('DefaultSession tests', () => {
   test('should create a no-session if storage is empty', () => {
