+++ src/gmp/session/__tests__/user-session-state.test.ts
@@ -3,19 +3,10 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import {describe, test, expect, testing} from '@gsa/testing';
+import {describe, test, expect} from '@gsa/testing';
 import date from 'gmp/models/date';
 import UserSessionState from 'gmp/session/user-session-state';
-
-const createStorage = (state?: Record<string, string>) => {
-  const store = {
-    state: {...state},
-    getItem: testing.fn(name => store.state[name] ?? null),
-    setItem: testing.fn((name, value) => (store.state[name] = String(value))),
-    removeItem: testing.fn(name => delete store.state[name]),
-  };
-  return store;
-};
+import {createStorage} from 'gmp/testing';
 
 describe('UserSessionState tests', () => {
   test('should create a session state', () => {
